Further separate EmailInvite from Project.
